feat: organize folder estructures
