+++ src/components/locale-toggler/index.tsx
@@ -10,7 +10,7 @@ import {
   DropdownMenuTrigger,
 } from '@/components/ui/dropdown-menu';
 
-import { Button } from '../../ui/button';
+import { Button } from '../ui/button';
 
 const locales = [
   {
