+++ src/modules/submit/components/done/index.tsx
@@ -3,7 +3,7 @@ import { Link } from 'react-router-dom';
 import { CheckCircle } from 'lucide-react';
 
 import { Button } from '@/components/ui/button';
-import { LINKS } from '@/constants';
+import { LINKS } from '@/config/constants';
 // import { useAuth0 } from '@auth0/auth0-react';
 
 export const Done = () => {
