+++ src/pages/calculator/contact/index.tsx
@@ -1,0 +1,5 @@
+import CalculatorContactScreen from '@/modules/calculator/screens/contact';
+
+export default function CalculatorContact() {
+  return <CalculatorContactScreen />;
+}
