+++ src/pages/calculator/index.tsx
