+++ src/pages/calculator/steps/index.tsx
@@ -1,5 +1,5 @@
 import CalculatorStepsScreen from '@/modules/calculator/screens/steps';
 
-export default function Calculator() {
+export default function CalculatorSteps() {
   return <CalculatorStepsScreen />;
 }
