+++ src/pages/index.ts
@@ -1,0 +1,7 @@
+export * from './calculator'
+export * from './dashboard'
+export * from './home'
+export * from './kyc'
+export * from './onboarding'
+export * from './profile'
+export * from './submit'
