+++ src/pages/onboarding/index.tsx
